Suppress docker-compose arg warnings.
If the user has not created a .env file, then they see warnings like
this when running `docker-compose`:

```
WARNING: The USER_ID variable is not set. Defaulting to a blank string.
WARNING: The GROUP_ID variable is not set. Defaulting to a blank string.
```

This fix provides defaults to docker-compose.yml to suppress this
warning.

See #2739